Conflicts:
	CGVFrontend/.classpath

+++ CGVFrontend/src/com/kosmo/cgv/web/MovieController.java
@@ -26,7 +26,7 @@ public class MovieController {
 	public String showMovieChart(Model model) throws Exception{
 		List<MovieDto> list = service.selectMovieList();
 		Map<String, Double> reserveRateMap = new HashMap<String, Double>();
-		Map<String, Integer> dDayMap = new HashMap<String, Integer>();
+		Map<String, Long> dDayMap = new HashMap<String, Long>();
 		Map<String, Integer> wishesMap = new HashMap<String, Integer>();
 		for(MovieDto movie: list) {
 			String movie_code = movie.getMovie_code();
@@ -47,11 +47,11 @@ public class MovieController {
 			double reserveRate = (seatCount/totalSeatCount)*100.0;
 			reserveRateMap.put(movie_code, reserveRate);
 			
-			java.sql.Date date = movie.getReleasedate();
-			java.sql.Date today = new java.sql.Date(System.currentTimeMillis());
+			java.sql.Date date = movie.getReleasedate();			
+			java.sql.Date today = new java.sql.Date(System.currentTimeMillis());			
 			long dateTime = date.getTime();
 			long todayTime = today.getTime();
-			int dayGap = (int)(dateTime - todayTime)/(1000*60*60*24);
+			long dayGap = (dateTime - todayTime)/(1000*60*60*24);			
 			dDayMap.put(movie_code, dayGap);
 			
 			int wishes = service.getWishCount(movie_code);
